Ensure pointers are kept alive during Windows syscalls
The expression `uintptr(unsafe.Pointer(&pcred))` gets an address of a pointer in order to share memory with a Windows syscall. However, Go may reallocate pointer addresses at any time, and if that happens during a syscall, well, that's catastrophic.

The Go compiler already knows that it should keep the uintptr result static during a call to `syscall.Syscall`, but not during a `syscall.Proc.Call`. This manually instructs the compiler to extend the same behavior to Go functions in this library that make syscalls via Procs.